:repeat: localy without fabric
